18 Authorization Using Policies

+++ app/Http/Controllers/IdeaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\IdeaRequest;
 use App\Models\Idea;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -38,7 +39,7 @@ class IdeaController extends Controller
 
     public function show(Idea $idea)
     {
-        $this->authorizeCurrentUserIdea($idea);
+        Gate::authorize('update', $idea);
 
         return view('ideas.show', [
             'idea' => $idea,
@@ -47,7 +48,7 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
-        $this->authorizeCurrentUserIdea($idea);
+        Gate::authorize('update', $idea);
 
         return view('ideas.edit', [
             'idea' => $idea,
@@ -56,7 +57,7 @@ class IdeaController extends Controller
 
     public function update(IdeaRequest $request, Idea $idea)
     {
-        $this->authorizeCurrentUserIdea($idea);
+        Gate::authorize('update', $idea);
 
         $validated = $request->validated();
 
@@ -69,15 +70,10 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
-        $this->authorizeCurrentUserIdea($idea);
+        Gate::authorize('update', $idea);
 
         $idea->delete();
 
         return redirect('/ideas');
-    }
-
-    private function authorizeCurrentUserIdea(Idea $idea): void
-    {
-        abort_unless($idea->user_id === auth()->id(), 403);
     }
 }
